+++ packages/client/wallets/aa/src/blockchain/wallets/FireblocksNCWallet.ts
@@ -35,7 +35,7 @@ export const FireblocksNCWallet = async ({
     passphrase,
     ncwData,
 }: FireblocksNCWWalletInput) => {
-    const localStorageRepository = new LocalStorageRepository(userIdentifierString(userIdentifier), projectId);
+    const localStorageRepository = new LocalStorageRepository(userIdentifier, projectId);
 
     let _walletId: string;
     let _deviceId: string;
@@ -211,14 +211,3 @@ const getDefaultAlgorithems = (): Set<TMPCAlgorithm> => {
     algorithms.add("MPC_CMP_ECDSA_SECP256K1");
     return algorithms;
 };
-
-function userIdentifierString(userIdentifier: UserIdentifier) {
-    switch (userIdentifier.type) {
-        case "email":
-            return `email:${userIdentifier.email}`;
-        case "phoneNumber":
-            return `phoneNumber:${userIdentifier.phoneNumber}`;
-        case "whiteLabel":
-            return `userId:${userIdentifier.userId}`;
-    }
-}
